Fixed issue with AutoLayout
When using AutoLayout setting the coach mark’s frame in viewDidLoad
does not guarantee that the view’s frames have been properly set.
Instead of passing in a CGRect we pass in the UIView itself and its
root view along with its parent view. We calculate the CGRect of the
view when the coach mark starts. This ensures that the view’s layout
has been finalized. The parentView and rootView are needed in order to
properly translate the view’s frame to the root view’s frame.